+++ pkg/queryfrontend/config.go
@@ -205,6 +205,9 @@ type Config struct {
 	DownstreamURL          string
 	ForwardHeaders         []string
 	NumShards              int
+	TenantHeader           string
+	DefaultTenant          string
+	TenantCertField        string
 }
 
 // QueryRangeConfig holds the config for query range tripperware.