+++ pkg/rules/manager_test.go
@@ -42,11 +42,16 @@ type nopAppender struct{}
 func (n nopAppender) Append(storage.SeriesRef, labels.Labels, int64, float64) (storage.SeriesRef, error) {
 	return 0, nil
 }
+
 func (n nopAppender) AppendExemplar(storage.SeriesRef, labels.Labels, exemplar.Exemplar) (storage.SeriesRef, error) {
 	return 0, nil
 }
 
 func (n nopAppender) AppendHistogram(ref storage.SeriesRef, l labels.Labels, t int64, h *histogram.Histogram, fh *histogram.FloatHistogram) (storage.SeriesRef, error) {
+	return 0, nil
+}
+
+func (n nopAppender) AppendCTZeroSample(ref storage.SeriesRef, l labels.Labels, t, ct int64) (storage.SeriesRef, error) {
 	return 0, nil
 }
 
@@ -381,7 +386,7 @@ groups:
 				return nil, nil
 			}
 		},
-		nil,
+		labels.EmptyLabels(),
 		"http://localhost",
 	)
 
@@ -439,7 +444,7 @@ groups:
 				}, nil
 			}
 		},
-		nil,
+		labels.EmptyLabels(),
 		"http://localhost",
 	)
 	thanosRuleMgr.Run()