+++ pkg/store/cache/factory.go
@@ -41,7 +41,7 @@ type IndexCacheConfig struct {
 func NewIndexCache(logger log.Logger, confContentYaml []byte, reg prometheus.Registerer) (IndexCache, error) {
 	level.Info(logger).Log("msg", "loading index cache configuration")
 	cacheConfig := &IndexCacheConfig{}
-	cacheMetrics := newCommonMetrics(reg)
+	cacheMetrics := NewCommonMetrics(reg)
 	if err := yaml.UnmarshalStrict(confContentYaml, cacheConfig); err != nil {
 		return nil, errors.Wrap(err, "parsing config YAML file")
 	}
@@ -78,6 +78,7 @@ func NewIndexCache(logger log.Logger, confContentYaml []byte, reg prometheus.Reg
 		return nil, errors.Wrap(err, fmt.Sprintf("create %s index cache", cacheConfig.Type))
 	}
 
+	cache = NewTracingIndexCache(string(cacheConfig.Type), cache)
 	if len(cacheConfig.EnabledItems) > 0 {
 		if err = ValidateEnabledItems(cacheConfig.EnabledItems); err != nil {
 			return nil, err